requirements.txt
flask==2.3.3
python-dotenv==1.0.0
neo4j==4.4.10
openai==1.3.5
flask-socketio==5.3.4
gunicorn==21.2.0
httpx==0.25.0
eventlet==0.33.3
python-engineio==4.5.1
python-socketio==5.8.0
werkzeug==2.3.7
